refactor: decouple local-first page state
